ref(openai)!: Move OpenAI AI integration to `@sentry/server-utils`
Move the OpenAI gen-ai instrumentation from `@sentry/core` to
`@sentry/server-utils/src/ai/openai`, consuming the shared `ai/core` base. The
`instrumentOpenAiClient`/`OPENAI_INTEGRATION_NAME`/etc. symbols are no longer
exported from `@sentry/core`; node/cloudflare/vercel-edge re-export them from
`@sentry/server-utils`, so the platform SDK public API is unchanged.

`providerSkip` stays in `@sentry/core` until the final cleanup PR.
