+++ packages/server-utils/src/ai/index.ts
@@ -1,0 +1,7 @@
+/**
+ * AI/gen-ai instrumentation logic for server runtimes.
+ *
+ * @module
+ */
+
+export { instrumentOpenAiClient } from './openai';
