+++ packages/server-utils/src/ai/openai/types.ts
