+++ packages/vercel-edge/src/index.ts
@@ -73,7 +73,6 @@ export {
   functionToStringIntegration,
   // eslint-disable-next-line typescript/no-deprecated
   inboundFiltersIntegration,
-  instrumentOpenAiClient,
   instrumentStateGraph,
   instrumentGoogleGenAIClient,
   instrumentAnthropicAiClient,
@@ -108,6 +107,7 @@ export {
   withStreamedSpan,
   spanStreamingIntegration,
 } from '@sentry/core';
+export { instrumentOpenAiClient } from '@sentry/server-utils/no-diagnostic-channels';
 
 export { VercelEdgeClient } from './client';
 export { getDefaultIntegrations, init } from './sdk';
